[WIP] More Vision work.
