+++ vision/google/cloud/vision/helpers.py
@@ -18,10 +18,7 @@ import six
 
 from google.gax.utils import protobuf
 
-from google.cloud.vision import meta
 
-
-@six.add_metaclass(meta.VisionHelpersMeta)
 class VisionHelpers(object):
     """A set of convenience methods to make the Vision GAPIC easier to use.
