+++ vision/requirements.txt
@@ -1,4 +1,3 @@
-googleapis-common-protos >=1.5.2, <2.0dev
-google-gax >=0.15.10, <0.16dev
-requests >= 2.13.0, < 3.0dev
+googleapis-common-protos >= 1.5.2, < 2.0dev
+google-gax >= 0.15.11, < 0.16dev
 six >= 1.10.0
